engine/requirements.txt
python-socketio >= 5.3.0
aiohttp == 3.7.4.post0
aiohttp_cors == 0.7.0
numpy
psutil == 5.8.0
scipy == 1.7.1
pynwb == 2.0.0
pyyaml
paramiko